move register_operator to Executor, common sdpa fn to utils
